# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Internship API",
    "description": "API for School Internship Management System - Track student internships, daily logs, attendance, and evaluations",
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:8000/api/v1", "description": "Local Development Server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "admin@school.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" },
                    "token_type": { "type": "string", "example": "bearer" },
                    "expires_in": { "type": "integer", "example": 3600 }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string", "enum": ["school_admin", "teacher", "student", "company_supervisor"] },
                  "school_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthenticated" }
        }
      }
    },
...
  }
}